put the shared struct in main
