+++ server.go
@@ -2,7 +2,6 @@ package main
 
 import (
   "io"
-  "sync"
   "bytes"
   "strings"
   "strconv"
@@ -10,51 +9,8 @@ import (
   "math/rand"
   "net/http"
   "encoding/json"
-  "github.com/syndtr/goleveldb/leveldb"
   "github.com/syndtr/goleveldb/leveldb/util"
 )
-
-// *** Shared App struct and methods ***
-
-type App struct {
-  db *leveldb.DB
-  mlock sync.Mutex
-  lock map[string]struct{}
-
-  // params
-  volumes []string
-  fallback string
-  replicas int
-  subvolumes int
-  softdelete bool
-}
-
-func (a *App) UnlockKey(key []byte) {
-  a.mlock.Lock()
-  delete(a.lock, string(key))
-  a.mlock.Unlock()
-}
-
-func (a *App) LockKey(key []byte) bool {
-  a.mlock.Lock()
-  defer a.mlock.Unlock()
-  if _, prs := a.lock[string(key)]; prs {
-    return false
-  }
-  a.lock[string(key)] = struct{}{}
-  return true
-}
-
-func (a *App) GetRecord(key []byte) Record {
-  data, err := a.db.Get(key, nil)
-  rec := Record{[]string{}, true}
-  if err != leveldb.ErrNotFound { rec = toRecord(data) }
-  return rec
-}
-
-func (a *App) PutRecord(key []byte, rec Record) bool {
-  return a.db.Put(key, fromRecord(rec), nil) == nil
-}
 
 // *** Master Server ***
